+++ wasm_apps/chronometer/src/main.rs
@@ -22,7 +22,7 @@ static mut APP_STATE: OnceCell<AppState> = OnceCell::new();
 #[no_mangle]
 pub fn init() -> () {
     let win_rect = guestlib::get_win_rect();
-    let fb_handle = guestlib::create_framebuffer(win_rect.w as usize, win_rect.h as usize);
+    let fb_handle = guestlib::create_framebuffer(win_rect.w, win_rect.h);
     let state = AppState { fb_handle };
     unsafe { APP_STATE.set(state).expect("App already initialized"); }
 }
